% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis.R
\name{clusterAnalysis}
\alias{clusterAnalysis}
\title{clusterAnalysis}
\usage{
clusterAnalysis(expres, genes, samples, sampleNames, comparisonName,
  anotPackage, my.symbols = NULL, outputDir, fileOfLinks, numClusters,
  rowDistance, colDistance, RowVals = TRUE, ColVals = TRUE, colorsSet,
  colsForGroups, escala, densityInfo = "none", cexForColumns, cexForRows,
  plotProfiles = FALSE, Title = "", csvType = NULL)
}
\arguments{
\item{expres}{Expresion oused to perform the clusters.}

\item{genes}{????}

\item{samples}{????}

\item{sampleNames}{Name of the samples of the study.}

\item{comparisonName}{Name of the comparison. To identify the output}

\item{anotPackage}{Annotation package.}

\item{my.symbols}{????}

\item{outputDir}{Path of the file created.}

\item{fileOfLinks}{Name of the links file.}

\item{numClusters}{????}

\item{rowDistance}{Row distance in the clusters.}

\item{colDistance}{Column distance in the clusters.}

\item{RowVals}{????}

\item{ColVals}{?????}

\item{colorsSet}{???}

\item{colsForGroups}{Colors of each group for the plots.}

\item{escala}{???}

\item{densityInfo}{???}

\item{cexForColumns}{Cex for the columns.}

\item{cexForRows}{Cex for the rows.}

\item{plotProfiles}{???}

\item{Title}{Title of the plot.}

\item{csvType}{Csv type of the output.}
}
\description{
Function to make the cluster analysis.
}
\examples{
 \dontrun{
for(i in 1:length(compName)) {
 matrices <- as.logical(apply(abs(as.matrix(cont.matrix[, wCont[[i]]])), 1, sum))
 s2clust <-  which(as.logical(apply(design[, matrices], 1, sum)))
 pvalType <- ifelse(adjMethod[i] == "none", "pvalues", "adj-pvalues")
 geneListFName <- paste("geneList", compName[i], pvalType, "LT", pValCutOff[i], "Rda", sep = ".")
 pal <- colorpanel(n = 32, low = "green", mid = "white", high = "magenta")
 load("./ResultsDir/geneList.Group1.pvalues.LT.0.01.Rda")
 clust <- BasicP::clusterAnalysis(expres = exprs.filtered,
                                  genes = geneList,
                                  samples = s2clust,
                                  sampleNames = as.character(targets$ShortName)[s2clust],
                                  comparisonName = "Compar 1",
                                  anotPackage = "org.Hs.eg",
                                  my.symbols = symbolsTable,
                                 outputDir = outputDir,
                                  fileOfLinks = linksFile,
                                  numClusters = 2,
                                  rowDistance = NULL,
                                  colDistance = NULL,
                                  RowVals = TRUE,
                                  ColVals = FALSE,
                                  escala = "row",
                                  colorsSet = pal,
                                  densityInfo = "density",
                                  colsForGroups = c(rep("pink", 5), rep("blue", 5)),
                                  cexForColumns = 0.8,
                                  cexForRows = 0.8,
                                  Title = paste("Compar 1 with", pvalType, "<", pValCutOff[i],
                                                ifelse(minLogFoldChange[i] == 0, "",
                                                       paste0("\\n and |logFC|>=",
                                                              minLogFoldChange[i]))),
                                  csvType = "csv2")
}
}
}
