% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOrLoadAnnotations.R
\name{createOrLoadAnnotations}
\alias{createOrLoadAnnotations}
\title{createOrLoadAnnotations}
\usage{
createOrLoadAnnotations(loadAnnotations = FALSE, chipPackAvailable,
  platformDesignPackAvailable, chipPackage, platformDesignPackage,
  outputDir, annotationsFileName, entrezTableFileName,
  symbolsTableFileName, controlsTableFileName)
}
\arguments{
\item{loadAnnotations}{FALSE by default. If TRUE the function loads the annotations files.}

\item{chipPackAvailable}{TRUE if there is a chip package available.}

\item{platformDesignPackAvailable}{TRUE if there is a platform design package available.}

\item{chipPackage}{Name of the chip package. Only if chipPackAvailable is TRUE.}

\item{platformDesignPackage}{Name of the platform design. Only if platformDesignPackAvailable is TRUE.}

\item{outputDir}{Path where the annotation will be stored.}

\item{annotationsFileName}{Name of the file for annotations.}

\item{entrezTableFileName}{Name of the file for Entrez genes.}

\item{symbolsTableFileName}{Name of the file for gene symbols ID.}

\item{controlsTableFileName}{Name of the file for controls.}
}
\value{
A list "anotacions" that contains the annotations, the Entrez genes and the gene symbols ID.
}
\description{
Function that creates the annotation needed, or if it is already done, loads the file created.
The annotation packages used need to be installed previously.
}
\examples{
\dontrun{
loadAnnotations <- FALSE
chipPackAvailable <- TRUE
platformDesignPackAvailable <- FALSE
chipPackage <- "hgu133a2"
platformDesignPackage <- NULL
outputDir <- "./ResultsDir"
annotationsFileName <- "Annotations"
entrezTableFileName <-"Entrezs.Rda"
symbolsTableFileName <-"Symbols.Rda"
controlsTableFileName <- "controls.Rda"


anotacions <- createOrLoadAnnotations (loadAnnotations= loadAnnotations,
chipPackAvailable = chipPackAvailable, platformDesignPackAvailable = platformDesignPackAvailable,
chipPackage = chipPackage, platformDesignPackage = platformDesignPackage,
outputDir = outputDir,annotationsFileName = annotationsFileName,
entrezTableFileName = entrezTableFileName, symbolsTableFileName = symbolsTableFileName,
 controlsTableFileName = controlsTableFileName)}
}
