% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGeneAnnotation.R
\name{doGeneAnnotation}
\alias{doGeneAnnotation}
\title{doGeneAnnotation}
\usage{
doGeneAnnotation(AnotList)
}
\arguments{
\item{AnotList}{List object that contains the parameters needed to carry out the analysis.}
}
\description{
Function that annotates a data set.
}
\examples{
\dontrun{
organisme <- "hsa"
AnotList <- list(fitMain = NULL, fitFileName = fitFileName, my.IDs = "entrezTable",
                 anotPackage = "org.Hs.eg", toHTML = TRUE, outputDir = outputDir,
                 anotFilename = "Annotations",
                 titleAnotations = "Annotations for all genes analyzed", specie = "homo sapiens",
                 info2show = c("Affymetrix", "EntrezGene", "GeneSymbol",
                               "GeneName", "KEGG", "GO"),
                 linksFile = linksFile, numGenesPerPage = NULL)

BasicP::doGeneAnnotation(AnotList)}
}
