% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGOAnalysis.R
\name{doGOAnalysis}
\alias{doGOAnalysis}
\title{doGOAnalysis}
\usage{
doGOAnalysis(GOPar)
}
\arguments{
\item{GOParsList}{List object that contains the parameters needed to carry out the analysis.}
}
\value{
GOResult
}
\description{
Function to interate with GOAnalysis
}
\examples{
\dontrun{
GOResult<-BasicP::GOAnalysis(fitMain = fitMain, whichContrasts = wCont,
comparison.Name = "Estudi", outputDir = outputDir, anotPackage = "org.Hs.eg",
my.IDs = entrezTable, addGeneNames = TRUE, fileOfLinks = linksFile, thrLogFC = 1,
cutoffMethod = "adjusted", P.Value.cutoff = rep(0.05, length(wCont)), pval = 0.01,
min.count = 3, ontologias = c("MF", "BP", "CC"), testDirections = c("over", "under"),
minNumGens = 0)
}
}
