% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doLmAnalysis.R
\name{doLmAnalysis}
\alias{doLmAnalysis}
\title{doLmAnalysis
Function to make the Linear model analysis. The unique parameter is lmPar list.}
\usage{
doLmAnalysis(lmPar)
}
\arguments{
\item{lmPar}{list object that contains the parameters needed to carry out the analysis.}
}
\description{
doLmAnalysis
Function to make the Linear model analysis. The unique parameter is lmPar list.
}
\examples{
\dontrun{
lmParsList <- list()
Estudi <- list(dades = NULL,
               expresFileName = "exprs.filtered.Rda",
               targets = targets,
               designMat = design,
               contMat = cont.matrix,
               whichContrasts = 1:ncol(cont.matrix),
               anotPack = NULL,
               outputDir = outputDir,
               ExpressionsAndTop = TRUE,
               showLmParams = FALSE,
               use.dupCorr = FALSE,
               block = NULL,
               nDups = 1,
               comparisonName = comparison,
               ENTREZIDs = "entrezTable",
               SYMBOLIDs = "symbolsTable",
               fileOfLinks = linksFile,
               fitFileName = fitFileName,
               csvType=csvType,
               rows2HTML = NULL,
               anotFilename = anotFileName
)

lmParsList <- add2parsList(lmParsList, Estudi)

for(ix in 1:length(lmParsList))
{
  fit.Main   <- BasicP::doLmAnalysis(lmParsList[ix])
}
}

}
