% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMultCompAnalysis.R
\name{doMultCompAnalysis}
\alias{doMultCompAnalysis}
\title{doMultCompAnalysis}
\usage{
doMultCompAnalysis(mcPar)
}
\arguments{
\item{mcPar}{List object that contains the parameters.}
}
\description{
Function for make the comparative analysis.
}
\examples{
\dontrun{
mcParsList <- list()
compName <- c("Group1", "Group2")
wCont <- 1:ncol(cont.matrix)
pValCutOff <- c(0.01, 0.01)
adjMethod <- c("none", "none")
minLogFoldChange <- c(1, 1)
for(i in 1:length(compName))
{
  pvalType <- ifelse(adjMethod[i] == "none", "p-values", "adj. p-values")
  mci <- list(fitMain = fitMain,
              fitFileName = fitFileName,
              whichContrasts = wCont[[i]],
              comparisonName = compName[i],
              titleText = paste("for", pvalType, "<", pValCutOff[i],
                                "and |logFC| >", minLogFoldChange[i]),
              anotPackage = "org.Hs.eg",
              my.symbols = symbolsTable,
              outputDir = outputDir,
              fileOfLinks = linksFile,
              multCompMethod = "separate",
              adjustMethod = adjMethod[i],
              selectionType = "any",
              P.Value.cutoff = pValCutOff[i],
              plotVenn = TRUE,
              colsVenn = NULL,
              vennColors= c("red", "yellow", "green", "blue", "pink"),
              cexVenn = 1,
              geneListFName = paste("geneList", compName[i], pvalType,
                                    "LT", pValCutOff[i], "Rda", sep = "."),
              minLogFC = minLogFoldChange[i],
              csvType = csvType)

  mcParsList <- add2parsList(mcParsList, mci)
}

for(ix in 1:length(mcParsList))
{
  geneList.MCi <- BasicP::doMultCompAnalysis(mcParsList[ix])
}
}
}
