% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{filterData}
\usage{
filterData(expres, controls = NULL, removeNAs = FALSE,
  entrezs = NULL, bySignal = TRUE, signalThr, grups = NULL,
  sigFun.Name = "filt.by.signal", sigThr.as.perc = TRUE,
  byVar = TRUE, variabilityThr, varFun.Name = "sdf",
  varThr.as.perc = TRUE, pairingFun.Name = NULL, targets,
  doReport = NULL, outputDir = NULL, filteringReportFName = NULL)
}
\arguments{
\item{expres}{Data set to be filtered.}

\item{controls}{Vector that indicates wich observations are controls.}

\item{removeNAs}{If TRUE the Na's observations will be deleted.}

\item{entrezs}{Name of the file for Entrez genes.}

\item{bySignal}{If TRUE the data is filtered by signal value.}

\item{signalThr}{Threshold for filter the data.}

\item{grups}{Name of the different groups of the data.}

\item{sigFun.Name}{????????}

\item{sigThr.as.perc}{???????????}

\item{byVar}{If TRUE the data is filtered by ?????}

\item{variabilityThr}{Threshold for filter the data.}

\item{varFun.Name}{????????}

\item{pairingFun.Name}{Name of the function to be imported to pair data}

\item{targets}{Name of the targets file.}

\item{doReport}{If TRUE a report is created.}

\item{outputDir}{Path of the file created.}

\item{filteringReportFName}{Name of the report file.}
}
\value{
Matrix object with te data filtered
}
\description{
Function that filter data and create a report file.
}
\examples{
\dontrun{
load("./ResultsDir/normalizedData.Rda")
repes <- duplicated(exprs(my.norm), MARGIN=1)
exprs(my.norm) <- exprs(my.norm)[!repes,]
eset_norm <- my.norm
load("./ResultsDir/controls.Rda")
removeNAs <- TRUE
load("./ResultsDir/Entrezs.Rda")
entrezs <- entrezTable
SignalFilter <- TRUE
signalThreshold <- 50
signalFilter.Function <- "filt.by.Signal"
signalThreshold.as.percentage <- TRUE
VarFilter <- TRUE 
variabilityThreshold <- 50
variability.Function <- "sdf"
variabilityThreshold.as.percentage <- TRUE
pairing.Function <- NULL
my.targets <-read.AnnotatedDataFrame("./celfiles/targets.txt", header = TRUE, row.names = 1)
doReport <- TRUE
outputDir <- "./ResultsDir"
FilteringReportFileName <- "FilteringReport.txt"
exprs.filtered <- filterData(expres = exprs(eset_norm),controls = names(controlsTable),
removeNAs = TRUE, entrezs = entrezs ,bySignal = SignalFilter,signalThr = signalThreshold,
grups = pData(eset_norm)$grupo, sigFun.Name = signalFilter.Function, 
sigThr.as.perc = signalThreshold.as.percentage, byVar = VarFilter, variabilityThr = variabilityThreshold,
varFun.Name = variability.Function, varThr.as.perc = variabilityThreshold.as.percentage,
pairingFun.Name = pairing.Function, targets = my.targets, doReport = doReport, outputDir = outputDir, 
filteringReportFName = FilteringReportFileName)}
}
