% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAnnotation.R
\name{GeneAnnotation.}
\alias{GeneAnnotation.}
\alias{GeneAnnotation}
\title{Function to create a table with annotations to provide information about elected genes.}
\usage{
GeneAnnotation(egIDs, anotPackage, toHTML = TRUE, outputDir,
  filename = "annotations",
  myTitle = "Annotations for all genes analyzed",
  specie = "Homo_sapiens", info2show = c("Affymetrix", "EntrezGene"),
  linksFile, maxGenes = NULL, ...)
}
\arguments{
\item{egIDs}{Entrez gene identifiers.}

\item{anotPackage}{Annotation package.}

\item{toHTML}{If TRUE a html file will be created.}

\item{outputDir}{Path of the file created.}

\item{filename}{Name of the file.}

\item{myTitle}{Title of ???}

\item{specie}{Specie}

\item{info2show}{Information that has to be shown}

\item{linksFile}{Name of the LinksFile.}

\item{maxGenes}{Maximum number of}
}
\value{
It returns the time that the process least.
}
\description{
Function to create a table with annotations to provide information about elected genes.
}
\examples{
\dontrun{
genes2annotate <- entrezs[unique(rownames(fitMain$p.value))]
genesAnnotated <-BasicP::GeneAnnotation(egIDs = genes2annotate, anotPackage = "org.Hs.eg",
toHTML = TRUE, outputDir = outputDir, filename = "Annotations",
myTitle = "Annotations for all genes analyzed", specie = "homo sapiens",
info2show = c( "Affymetrix", "EntrezGene", "GeneSymbol", "GeneName", "KEGG", "GO"),
linksFile = linksFile, maxGenes = NULL)
}
}
