% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoAnalysis.R
\name{GOAnalysis}
\alias{GOAnalysis}
\title{Function that perform enrichment analysis based on the GO for one gene list and all ontologies with the GOstats package.}
\usage{
GOAnalysis(fitMain, whichContrasts, comparison.Name, outputDir,
  anotPackage, my.IDs, addGeneNames = TRUE, fileOfLinks,
  thrLogFC = NULL, cutoffMethod = c("adjusted", "unadjusted"),
  P.Value.cutoff = rep(0.05, length(whichContrasts)), pval = 0.05,
  min.count = 3, ontologias = c("MF", "BP", "CC"),
  testDirections = c("over", "under"), minNumGens = 0)
}
\arguments{
\item{fitMain}{Object generated by the lmFit function.}

\item{whichContrasts}{Describes the lmFit contrasts whose associated ids will be analyzed.}

\item{comparison.Name}{Name of the comparison. To identify the output.}

\item{outputDir}{Path of the file created.}

\item{anotPackage}{Annotation package.}

\item{my.IDs}{Gene identifiers to be analyze. If NULL all provided by the lmFit object will be analyzed.}

\item{addGeneNames}{Vector of the gene names.}

\item{fileOfLinks}{Name of the links file.}

\item{thrLogFC}{Threshold for the logarithm Fold Change}

\item{cutoffMethod}{Method used to select the genes to be analyzed.}

\item{P.Value.cutoff}{Maximum value for the p.value.}

\item{pval}{Pvalue for the enrichment analysis test.}

\item{min.count}{Minimum number of categories for a result to be considered.}

\item{ontologias}{Ontologies to be used in the analysis.}

\item{testDirections}{To decide if the analysis has to be applied to genes upregulated, downregulated or both.}

\item{minNumGens}{Mimimum number of genes.}
}
\description{
Function that perform enrichment analysis based on the GO for one gene list and all ontologies with the GOstats package.
}
\examples{
\dontrun{
library(GOstats)
GOResult<-BasicP::GOAnalysis(fitMain = fitMain, whichContrasts = wCont,
comparison.Name = "Estudi", outputDir = outputDir, anotPackage = "org.Hs.eg",
my.IDs = entrezTable, addGeneNames = TRUE, fileOfLinks = linksFile, thrLogFC = 1,
cutoffMethod = "adjusted", P.Value.cutoff = rep(0.05, length(wCont)), pval = 0.01,
min.count = 3, ontologias = c("MF", "BP", "CC"), testDirections = c("over", "under"),
minNumGens = 0)
}
}
