% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGAnalysis.R
\name{KEGGAnalysis}
\alias{KEGGAnalysis}
\title{KEGGAnalysis}
\usage{
KEGGAnalysis(fitMain, whichContrasts, comparison.Name, outputDir,
  anotPackage, organisme, my.IDs, addGeneNames = TRUE, fileOfLinks,
  cutoffMethod = c("adjusted", "unadjusted"),
  P.Value.cutoff = rep(0.05, length(whichContrasts)), pval = 0.05,
  thrLogFC = NULL, minNumGens = 0)
}
\arguments{
\item{fitMain}{????}

\item{whichContrasts}{????}

\item{comparison.Name}{????}

\item{outputDir}{????}

\item{anotPackage}{????}

\item{organisme}{????}

\item{my.IDs}{????}

\item{addGeneNames}{????}

\item{fileOfLinks}{????}

\item{cutoffMethod}{????}

\item{P.Value.cutoff}{????}

\item{pval}{????}

\item{thrLogFC}{????}

\item{minNumGens}{????}
}
\description{
Function to make KEGG analysis.
}
\examples{
\dontrun{
load("./ResultsDir/fit.Rda")
wCont <- 1:3
comparison.Name <-"Estudi"
outputDir <- "./ResultsDir"
orgPackage <-"org.Hs.eg"
organisme <-"hsa"
load("./ResultsDir/Entrezs.Rda")
addGeneNames <- TRUE
linksFileName <- "Links.txt"
cutoffMethod <-"unadjusted"
pval <- 0.05
minNumGens <-0
runMulticore <- 0

KEGGResult <- BasicP::KEGGAnalysis(fitMain = fit.main, whichContrasts = wCont,
comparison.Name = comparison.Name, outputDir = outputDir, anotPackage = orgPackage,
organisme = organisme, my.IDs = entrezTable, addGeneNames = addGeneNames,
fileOfLinks = linksFileName, cutoffMethod = cutoffMethod,
P.Value.cutoff = rep(0.05, length(wCont)), pval = pval,thrLogFC = 1,
minNumGens = minNumGens)
}
}
