% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmAnalysis.R
\name{lmAnalysis}
\alias{lmAnalysis}
\title{Function to make the linear model analysis.}
\usage{
lmAnalysis(exprs.filtered, design, cont.matrix,
  contrasts2test = 1:ncol(cont.matrix), anotPackage,
  Expressions_And_Top = TRUE, showParams = FALSE,
  use.dupCorr = FALSE, block = NULL, nDups = 1, comparison = "",
  outputDir = ".", ENTREZIDs = NULL, SYMBOLIDs = NULL, linksFile,
  fitFileName, csvType, rows2HTML = NULL, anotFileName)
}
\arguments{
\item{exprs.filtered}{Dataset of filtered and normalized data.}

\item{design}{Design matrix.}

\item{cont.matrix}{Contrast matrix.}

\item{contrasts2test}{Numeric vector that contains the number of the columns for the contrasts to analyze.}

\item{anotPackage}{Annotation package.}

\item{Expressions_And_Top}{If is TRUE it will do "Expresions" and the toptable.}

\item{showParams}{If FALSe it won't show the lmFit parameters in the ExpressionsAndToptable.}

\item{use.dupCorr}{parameter for correlated analysis to be passed to limma}

\item{block}{parameter for correlated analysis to be passed to limma}

\item{nDups}{parameter for correlated analysis to be passed to limma}

\item{comparison}{Name of the comparations in all contrasts.}

\item{outputDir}{Path of the file created.}

\item{ENTREZIDs}{Name of the file for Entrez genes.}

\item{SYMBOLIDs}{Name of the file for gene Symbols .}

\item{linksFile}{Name of the file linksFile.}

\item{fitFileName}{Name of file containing object resulting from lmAnalysis}

\item{csvType}{type of csv to store the data.}

\item{rows2HTML}{How many lines have to be written to HTML output}

\item{anotFileName}{Name where annotations have to be saved.}
}
\description{
Function to make the linear model analysis.
}
\examples{
\dontrun{
load("./ResultsDir/exprs.filtered.Rda")
contrasts2test <- 1:ncol(cont.matrix)
anotPackage = NULL
comparison =  "Estudi"
outputDir = "./ResultsDir"
ENTREZIDs = "entrezTable"
SYMBOLIDs = "symbolsTable"
linksFile = "Links.txt"
fitFileName = "fit.Rda"
csvType= "csv"
rows2HTML= NULL
anotFileName <- "Annotations"
runMulticore = 0
toTIFF= FALSE
fitMain <- BasicP::lmAnalysis(exprs.filtered = exprs.filtered, design = design,
cont.matrix = cont.matrix, contrasts2test = contrasts2test, anotPackage = anotPackage,
outputDir = outputDir, comparison = comparison, Expressions_And_Top = TRUE ,
showParams = FALSE , use.dupCorr = FALSE, block = NULL, nDups = 1 , ENTREZIDs = ENTREZIDs,
SYMBOLIDs = SYMBOLIDs, linksFile = linksFile,fitFileName = fitFileName , csvType=csvType,
rows2HTML = NULL, anotFileName = anotFileName)
}
}
