% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleComp.R
\name{multipleComp}
\alias{multipleComp}
\title{multipleComp
Function to make multiple comparations.}
\usage{
multipleComp(fitMain, whichContrasts, comparisonName, titleText, outputDir,
  anotPackage, my.symbols = NULL, linksFile,
  multCompMethod = "separate", adjustMethod = "none",
  selectionType = c("any", "all", "anyDown", "anyUp", "allDown",
  "allUp"), P.Value.cutoff = 0.05, plotVenn, colsVenn = NULL,
  vennColors, cexVenn = 1, geneListFName = paste("geneList",
  comparisonName, "Rda", sep = "."), csvType = NULL, minLFC = 0)
}
\arguments{
\item{fitMain}{Object resulting from lmFit estimation}

\item{whichContrasts}{Vector that determie the contrast to do.}

\item{comparisonName}{Name of the comparisons.}

\item{titleText}{Main title for the multiple comparison analysis}

\item{outputDir}{Path of the file created.}

\item{anotPackage}{Annotation package.}

\item{my.symbols}{List of gene symbols to be used to annotate he output.}

\item{linksFile}{Name of the LinksFile.}

\item{multCompMethod}{Method to make the comparation.}

\item{adjustMethod}{Method to apply in the adjust.}

\item{selectionType}{Selection method to be used.}

\item{P.Value.cutoff}{Cutoff value for p.value.}

\item{plotVenn}{Boolean variable to decide if plot diagram has to be used.}

\item{colsVenn}{Columnns for the Venn diagram}

\item{vennColors}{Columns for the Venn diagram}

\item{cexVenn}{Expansion coefficient for Venn diagram labels}

\item{geneListFName}{Name of file containg gene list.}

\item{csvType}{Type of csv file to be written.}

\item{minLFC}{minimum log Fold change for selecting genes.}
}
\description{
multipleComp
Function to make multiple comparations.
}
\examples{
\dontrun{
compName <- c("Group1")
wCont <- 1:3
pValCutOff <- c(0.01)
adjMethod <- c("none")
minLogFoldChange <- c(1)
load("./ResultsDir/Symbols.Rda")
pvalType <- ifelse(adjMethod == "none", "p-values", "adj. p-values")
titleText <- paste("for", pvalType, "<", pValCutOff, "and |logFC| >", minLogFoldChange)
geneListFName <- paste("geneList", compName, pvalType ,"LT", pValCutOff, "Rda", sep = ".")

geneList <- BasicP::multipleComp(fitMain = fitMain, whichContrasts = wCont,
                                 comparisonName = compName, titleText = titleText,
                                 outputDir = outputDir, anotPackage = "org.Hs.eg",
                                 my.symbols = symbolsTable, linksFile = linksFile,
                                 multCompMethod = "separate",
                                 adjustMethod = adjMethod, selectionType = "any",
                                 P.Value.cutoff = pValCutOff, plotVenn = TRUE, colsVenn = NULL,
                                 vennColors = c("red", "yellow", "green", "blue", "pink"),
                                 cexVenn = 1, geneListFName = geneListFName, csvType = csvType,
                                 minLFC = minLogFoldChange)
}
}
