% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normplots2File.R
\name{normplots2File}
\alias{normplots2File}
\title{normplots2File}
\usage{
normplots2File(my.data, sampleNames, my.colors, my.groups,
  my.method = "average", my.cex = 0.7, posText = 4, dim3 = FALSE,
  fileName, outputDir, PCAPlots = TRUE, csv, lFile = NULL)
}
\arguments{
\item{my.data}{Normalized data.}

\item{sampleNames}{Names of the samples of the normalized data.}

\item{my.colors}{Colores used in the plots.}

\item{my.groups}{Name of the groups in the plots}

\item{my.method}{Method used to make the clusters.}

\item{my.cex}{Cex used in the plots.}

\item{posText}{Position of the text.}

\item{dim3}{If TRUE the PCA plot is in 3D.}

\item{fileName}{Name of the file created.}

\item{outputDir}{Path of the file created.}

\item{PCAPlots}{If TRUE PCA plots will be created.}

\item{csv}{Indicates the file type.}

\item{lFile}{Name of the links file.}
}
\description{
normplots2File plots the result of the normalized data and creates a file.
}
\examples{
\dontrun{
load("./ResultsDir/normalizedData.Rda")
repes <- duplicated(exprs(my.norm), MARGIN=1)
exprs(my.norm) <- exprs(my.norm)[!repes,]
eset_norm <- my.norm
my.colors <- rainbow(length(sampleNames(eset_norm)))
my.names <- pData(eset_norm)$ShortName
myCex<- 0.8
dim3 <- FALSE
fileName <- "NormalizedPlots.pdf"
outputDir <- "./ResultsDir"
PCAPlots <- TRUE
csv <- "csv2"
normplots2File(my.data = eset_norm, sampleNames = my.names, my.colors = my.colors,
my.groups = pData(eset_norm)$Group, my.method = "average",my.cex = myCex ,
posText = 2, dim3 = FALSE,fileName = fileName, outputDir = outputDir,PCAPlots = TRUE,
csv = fileType)}
}
