% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOrLoad.RawData.R
\name{readOrLoad.RawData}
\alias{readOrLoad.RawData}
\title{readOrLoad.RawData}
\usage{
readOrLoad.RawData(readCELS, phenoDat, fileNames, dataFName, outputDir,
  exonSt = FALSE, cdf = NULL)
}
\arguments{
\item{readCELS}{If TRUE it reads and stores de novo the raw data.}

\item{phenoDat}{Name of the file that contains phenodata information.}

\item{fileNames}{List of the file names to import.}

\item{dataFName}{Name of the file that stores raw data.}

\item{outputDir}{Path where the my.raw will be stored.}

\item{exonSt}{Default value FALSE. If TRUE the raw data is exon array.}

\item{cdf}{name of "cel definition file" package}
}
\value{
my.raw expresion set object with the raw data.
}
\description{
Function that reads the raw data or if the eset is already created it is loaded.
}
\examples{
\dontrun{
require(affy)
require(oligo) #ExonStudy
readCELS <- TRUE
my.targets <- "./celfiles/targets.txt"
targets<- read.table("./celfiles/targets.txt", head=TRUE, sep="\\t", row.names = 1)
my.fileNames <-paste("./celfiles/",rownames(targets),sep="")
rawDataFileName <- "rawData.Rda"
my.outputDir <- "."
isExonStudy <- FALSE
orgPackage <- "org.Hs.eg"

rawData <- readOrLoad.RawData(readCELS = readCELS, phenoDat = my.targets,
fileNames = my.fileNames, dataFName =rawDataFileName,outputDir = my.outputDir,
exonSt = isExonStudy)}
}
