% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{saveData}
\usage{
saveData(expres, expresNames = colnames(expres), expres.csv.FileName,
  csvType, description = "Normalized Values", anotPackage,
  SYMBOL = "SYMBOL", symbolsVector = NULL, expres.bin.FileName,
  linksFile, outputDir)
}
\arguments{
\item{expres}{Object to be saved.}

\item{expresNames}{Names of the samples of expression.}

\item{expres.csv.FileName}{Name of the csv file.}

\item{csvType}{Csv type.}

\item{description}{Description of the file for Linksfile.}

\item{anotPackage}{Package of annotations.}

\item{SYMBOL}{????????}

\item{symbolsVector}{Name of the symbols table.}

\item{expres.bin.FileName}{File name of the filtered data.}

\item{linksFile}{Character string that indicates the path to the txt file.}

\item{outputDir}{Path of the file created.}
}
\description{
Function that saves a R object in files.
}
\examples{
\dontrun{
load("./ResultsDir/normalizedData.Rda")
repes <- duplicated(exprs(my.norm), MARGIN=1)
exprs(my.norm) <- exprs(my.norm)[!repes,]
eset_norm <- my.norm
normalized.all.FileName <- "normalized.all"
fileType <-"csv2"
symbolsTable <- load("./ResultsDir/Symbols.Rda")
expres.all.FileName <- "expres.Rda"
linksFileName <- "Links.txt"
outputDir <- "./ResultsDir"
saveData(expres = exprs(eset_norm), expres.csv.FileName = normalized.all.FileName,
csvType=fileType, description = "Normalized values for all genes", anotPackage = NULL,
symbolsVector = symbolsTable, SYMBOL = "SYMBOL", expres.bin.FileName = expres.all.FileName,
linksFile = linksFileName, outputDir = outputDir)}
}
